Fix rendering and improve style.
